hpsa: add heartbeat sysfs host attribute
Patchwork-id: 40325
O-Subject: [RHEL6.2 PATCH]  hpsa: add heartbeat sysfs host attribute
Bugzilla: 730027
RH-Acked-by: Rob Evers <revers@redhat.com>

This is for bz#730027

Upstream post http://lkml.org/lkml/2011/8/12/270

The cciss driver had a CCISS_HEARTBEAT ioctl which was not implemented in hpsa.
This ioctl returned a counter from a register on the Smart Array which the
firmware would periodically update.  It can be used to detect certain kinds
of faults (e.g. controller lockup) by noticing when the value remains constant
for more than a second or two.