cifs: fix possible memory corruption in CIFSFindNext
Patchwork-id: 40333
O-Subject: [RHEL6 PATCH] BZ#730354: cifs: fix possible memory corruption in CIFSFindNext
Bugzilla: 730354
CVE: CVE-2011-3191
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Eugene Teo <eugene@redhat.com>

The name_len variable in CIFSFindNext is a signed int that gets set to
the resume_name_len in the cifs_search_info. The resume_name_len however
is unsigned and for some infolevels is populated directly from a 32 bit
value sent by the server.

If the server sends a very large value for this, then that value could
look negative when converted to a signed int. That would make that
value pass the PATH_MAX check later in CIFSFindNext. The name_len would
then be used as a length value for a memcpy. It would then be treated
as unsigned again, and the memcpy scribbles over a ton of memory.

Fix this by making the name_len an unsigned value in CIFSFindNext.

Cc: <stable@kernel.org>
Reported-by: Darren Lavender <dcl@hppine99.gbr.hp.com>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>