qdio: 2nd stage retry on SIGA-W busy conditions
Patchwork-id: 40356
O-Subject: [RHEL6.2 PATCH] [s390x]qdio: 2nd stage retry on SIGA-W busy conditions
Bugzilla: 732708

Description
-----------
HiperSockets device recoveries under heavy system load.

The SIGA-W might return with the busy bit set which means the device
was blocked.  The busy loop which retries the SIGA-W for 100us might
not be long enough when running under a heavily loaded z/VM.

Extend the retry mechanism by adding a longer second stage which
retries the SIGA-W for up to 10s.  In difference to the first retry
loop the second stage is using mdelay to stop the cpu between the
retries and thereby avoid additional preassure in on the hypervisor.
If the second stage retry is successfull a device reset is avoided.

Bugzilla
--------
BZ 732708
https://bugzilla.redhat.com/show_bug.cgi?id=732708

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 3.1
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=be8d97a540cec5d272b1e08e27741a3c3aa38f66

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>