jbd[2]: Use WRITE_SYNC_PLUG in journal_commit_transaction
Patchwork-id: 40463
O-Subject: [rhel6 patch] jbd[2]: Use WRITE_SYNC_PLUG in journal_commit_transaction
Bugzilla: 720918
RH-Acked-by: Christoph Hellwig <chellwig@redhat.com>
RH-Acked-by: Josef Bacik <josef@redhat.com>
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

Hi,

Tao Ma posted this patch, which allows for better merging of journal
I/O.  He showed a very simple workload, just a dd of 1MB followed by the
write-back of that data, where each journal buffer was sent to disk
individually, instead of merging adjacent I/Os.  In practice, I have
been unable to see this affect a real workload (but I only monitored
kernel compiling).  In theory, however, this may very well be
problematic for some workloads.  The patch was posted upstream, but it
was not merged as the on-stack plugging basically accomplishes the same
thing.  So, without the patch, we see a lot of individual journal I/Os
sent to disk.  After the patch, there is merging of requests as
expected.

This addresses bug 720918.  Comments, as always, are appreciated.

Cheers,
Jeff

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>