fcoe: Rearrange fcoe port and NPIV port cleanup
Patchwork-id: 40505
O-Subject: [RHEL6.2 PATCH 01/14] fcoe: Rearrange fcoe port and NPIV port cleanup
Bugzilla: 734961
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>
RH-Acked-by: Rob Evers <revers@redhat.com>

From: Neerav Parikh <Neerav.Parikh@intel.com>

When NPIV port destroy handler is called it does not do all the cleanup
required for the given NPIV port. This was happening as some of the
lport cleanup moved to fcoe_interface_cleanup() routine, which is not
called as part of the vport delete process.

This patch rearranges the sequence in which the fcoe_if_destory() and
fcoe_interface_cleanup() functions are being called from various places
in the code. It now matches the sequence they are constructed during the
create process for both N_Port as well as NPIV port.

[ Notes for RHEL6.2:
This upstream patch was backported for RHEL6.2 for:
1. dev_uc/mc_ kAPI name change
2. queue_work -> schedule_work

-- yi ]

Tested-by: Ross Brattain <ross.b.brattain@intel.com>
Signed-off-by: Neerav Parikh <Neerav.Parikh@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>
Signed-off-by: James Bottomley <JBottomley@Parallels.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>