Restore max mqueue message size to its previous RHEL 5 value
Patchwork-id: 40517
O-Subject: Re: [PATCH][RHEL 6] Restore max mqueue message size to its previous
	RHEL 5 value
Bugzilla: 730632
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>
RH-Acked-by: Anton Arapov <Anton@redhat.com>

On 09/07/2011 02:12 PM, Laszlo Ersek wrote:
> On 09/07/11 00:18, Doug Ledford wrote:
>> It appears to me that the logic in mq_attr_ok() is
>> faulty in regards to validating input, and it actually places several
>> restrictions on the input that aren't spelled out anywhere else (like
>> the size of the msg_max * msg_sizemax passed in can not wrap a ulong
>> counter).
>
> sizeof(long unsigned) should equal sizeof(size_t) wherever we try to
> build the kernel, I think. Therefore the idea seems correct; it tries to
> check if we can express the size of the memory block to be allocated
> later (?).

Actually, that's not the purpose.  The check for the size of the overall
message queue being less than ULONG_MAX is because we have to account
the size of the allocated queue against the application, so the size
must fit inside the data type we have in the task struct for accounting.
 My comment about it being faulty in regards to validating input are
that it doesn't check the hard limits of some items and enforce them
even on CAP_SYS_RESOURCE enabled apps.  If we have a hard limit, it
should be checked on CAP_SYS_RESOURCE apps, and the soft limit should be
checked on non CAP_SYS_RESOURCE apps.

In addition to making sure that this specific queue doesn't overflow the
accounting variable we have for it, in mqueue_get_inode (which is where
we actually allocate the queue and attach it to an inode that is then
attached to the process) we perform another check to make sure that the
overall queue size, plus the queue size of any other queues already
attached to this process, do not overflow the counter (see lines 149-158
in ipc/mqueue.c).

So, the old setting of INT_MAX, for all intents and purposes, is a
useless maximum.  Unless you are setting the queue depth to 1, INT_MAX +
msg_header_size * queue_depth will exceed ULONG_MAX and not be usable.
I've written a new patch that improves the documentation of the maximum
values I selected (partly based upon feedback from Jeremy West in
regards to our customer's actual needs, partly based upon POSIX
requirements), improves enforcement of the values, adds a minimum and
maximum bound on the number of queues (which doesn't directly map to
POSIX since they specify number of queues on a per process basis and we
enforce them on a per namespace basis), and should solve the customer's
problem.

v2: update defines and improve documentation
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>