Revert megaraid_sas: Driver only report tape drive, JBOD and logic drives
Patchwork-id: 40540
O-Subject: [RHEL6.2 PATCH] Revert megaraid_sas:  Driver only report tape drive,
	JBOD and logic drives
Bugzilla: 736667
RH-Acked-by: Rob Evers <revers@redhat.com>

resolves bz#736667

Revert "[scsi] megaraid_sas: Driver only report tape drive, JBOD and logic drives"
This reverts commit 466851e0bf7ff057723a21195673be5b74cab304.

The original patch was added as a workaround for a warning seen while boot in 619422
and a panic in 703087. Now, because the upstream patch 'enclosure: fix error path
to actually return ERR_PTR() on error' is added to kernel -171, the original patch
is no longer needed. The downside of that workaround is that the driver returns
-ENXIO for enclosures, cdroms..., so this revert makes this all accessible again.
(The root of the problem was a certain enclosure - the enclosure reported
duplicate entries in the element names field).

Brew taskID=3616874

Tomas

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>