pdpte registers are not flushed when PGD entry is changed in x86 PAE mode
Patchwork-id: 40542
O-Subject: [RHEL6.2 PATCH 3/3] [mm] pdpte registers are not flushed when PGD
	entry is changed in x86 PAE mode
Bugzilla: 691310
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Andrea Arcangeli <aarcange@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Johannes Weiner <jweiner@redhat.com>

From: Larry Woodman <lwoodman@redhat.com>

A massive numbers of minor page faults happen when executing SPECjbb2005. As a
result, SPECjbb2005 hangs up a few seconds to 1 hour. The page faults occur
even though the virtual memory addresses in question are already mapped to
physical memory correctly.

This phenomenon occurs on only RHEL6 x86 (PAE) kernel, and customer has not
seen it on RHEL5 so far.

The details of this problem are as follows.

When executing, it sometimes hangs up for a long time as below messages.

-------------
:
:
116.160: [GC 406278K->193826K(2071552K), 0.0117750 secs]
116.342: [GC 406626K->194122K(2071616K), 0.0118000 secs]
116.522: [GC

# SPECjbb2005 hangs up showing here.
--------------------

After a while, SPECjbb2005 restart to execute and finishes with no error. As a
result of several reproduction tests, customer found that it takes from five
seconds to an hour to be recovered from hanging up. In addition, this behaviour
sometimes disappears by executing other unrelated command.

When this phenomenon happens, customer can observe a large number of page
faults
on the java process.

Here is the output of /proc/<pid>/stat of the java process. The minor fault
value, the tenth field of this file, is normally about one million after
SPECjbb2005 on their test environment like the followings:

------------------
1744 (java) S 1742 1742 1721 34816 1742 4202496 968826 0 35834 0 218734 ...
------------------

However, when this problem happens, this value increases sharply.

----
2065 (java) S 2063 2063 1721 34816 2063 4202496 157573157 0 1 0 .....
...
2065 (java) S 2063 2063 1721 34816 2063 4202496 231388697 0 1 0 ......
...
2065 (java) S 2063 2063 1721 34816 2063 4202496 438851940 0 1 0 ....
...
2065 (java) S 2063 2063 1721 34816 2063 4202496 524209252 0 1 0 ...
....
-----

The attached upstream patch fixes this problem and BZ691310:

commit 4981d01eada5354d81c8929d5b2836829ba3df7b
Author: Shaohua Li <shaohua.li@intel.com>
Date:   Wed Mar 16 11:37:29 2011 +0800

    x86: Flush TLB if PGD entry is changed in i386 PAE mode

    According to intel CPU manual, every time PGD entry is changed in
i386 PAE
    mode, we need do a full TLB flush. Current code follows this and
there is
    comment for this too in the code.

    But current code misses the multi-threaded case. A changed page table
    might be used by several CPUs, every such CPU should flush TLB. Usually
    this isn't a problem, because we prepopulate all PGD entries at process
    fork. But when the process does munmap and follows new mmap, this issue
    will be triggered.

    When it happens, some CPUs keep doing page faults:

      http://marc.info/?l=linux-kernel&m=129915020508238&w=2

    Reported-by: Yasunori Goto<y-goto@jp.fujitsu.com>
    Tested-by: Yasunori Goto<y-goto@jp.fujitsu.com>
    Reviewed-by: Rik van Riel <riel@redhat.com>
    Signed-off-by: Shaohua Li<shaohua.li@intel.com>
    Cc: Mallick Asit K <asit.k.mallick@intel.com>
    Cc: Linus Torvalds <torvalds@linux-foundation.org>
    Cc: Andrew Morton <akpm@linux-foundation.org>
    Cc: linux-mm <linux-mm@kvack.org>
    Cc: stable <stable@kernel.org>
    LKML-Reference: <1300246649.2337.95.camel@sli10-conroe>
    Signed-off-by: Ingo Molnar <mingo@elte.hu>
--------------------------------------------------------------------------------------------

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>