bnx2fc: Do not reuse the fcoe connection id immediately
Patchwork-id: 40577
O-Subject: [RHEL6.2 PATCH 18/36] bnx2fc: Do not reuse the fcoe connection id
	immediately
Bugzilla: 727304
RH-Acked-by: Rob Evers <revers@redhat.com>

From: Bhanu Prakash Gollapudi <bprakash@broadcom.com>

Submitted to scsi-misc-2.6, yet to be accepted.
http://marc.info/?l=linux-scsi&m=131250476110539&w=2

CFC_DELETE is issued 2 secs after CONN_TERM is completed. If the session is
uploaded and offloaded immediately, it has to wait for the connection id to be
available.

Signed-off-by: Bhanu Prakash Gollapudi <bprakash@broadcom.com>