libfcoe: fix compilation when fcoe.ko is not used
Patchwork-id: 40591
O-Subject: [RHEL6.2 PATCH 36/36] libfcoe: fix compilation when fcoe.ko is not
	used.
Bugzilla: 727304
RH-Acked-by: Rob Evers <revers@redhat.com>

From: Mike Christie <mchristi@redhat.com>

If CONFIG_FCOE is not set, like on zfcp, then it will
result in a build failure because ndo_fcoe_get_wwn on
the net_device_ops is there.

Upstream this is fixed with this patch:
http://patchwork.ozlabs.org/patch/111838/

We cannot do that in rhel, because CONFIG_LIBFCOE is
set for zfcp and that would add the ndo_fcoe_get_wwn
field to the net_device_ops structs (it was not
defined before because the ifdefs around it in the
the original code was just for CONFIG_FCOE), which
breaks KABI.

This patch just adds a CONFIG_FCOE around the
ndo_fcoe_get_wwn use in the libfcoe module.

It has been compile tested here:
https://brewweb.devel.redhat.com/taskinfo?taskID=3632580

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>