intel_idle: Fix mismerge
Patchwork-id: 40609
O-Subject: [RHEL6.2 PATCH V2] 733730: intel_idle: Fix mismerge
Bugzilla: 733730
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Myron Stowe <mstowe@redhat.com>

rhbz: #733730

A patch mismerge could result in incorrect assumptions about the C states
in which the local apic timer was reliable. This appears to have caused
panics on boot on a small number of large Westmere machines, such as those
used in 733730. This corrects the patch to its upstream state.