e1000: don't enable dma receives until after dma address has been setup
Patchwork-id: 40644
O-Subject: [RHEL6.2 PATCH] e1000: don't enable dma receives until after dma address has been setup
Bugzilla: 703357
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Andy Gospodarek <gospo@redhat.com>

Resolves RHBZ 703357

This is a backport of a patch created by Andy Gospodarek and posted upstream
to the netdev mailing list.

Doing an 'ifconfig ethN down' followed by an 'ifconfig ethN up' on a qemu-kvm
guest system configured with an e1000 NIC can result in an 'unable to handle
kernel paging request at 0000000100000000' or 'bad page map in process ...'
or something similar.

These are the end result of the qemu-kvm's e1000 hardware emulator writing a
E1000_RXD_STAT_DD to the status' of a 4096-byte dma page of e1000_rx_desc
structures, AFTER the dma page has been freed.

And this use after free comes about from the following:

  A. An 'ifconfig ethN down' gets us into e1000_close(), which through a
     number of subfunctions results in:
	1. E1000_RCTL_EN being cleared in RCTL register.  [e1000_down()]
	2. dma_free_coherent() being called.  [e1000_free_rx_resources()]

  B. An 'ifconfig ethN up' gets us into e1000_open(), which through a number
     of subfunctions results in:
	1. dma_alloc_coherent() being called.  [e1000_setup_rx_resources()]
	2. E1000_RCTL_EN being set in RCTL register.  [e1000_setup_rctl()]
	3. E1000_RCTL_EN being cleared in RCTL register.
	   [e1000_configure_rx()]
	4. RDLEN, RDBAH and RDBAL registers being set to reflect the dma page
	   allocated in step B.1.  [e1000_configure_rx()]
	5. E1000_RCTL_EN being set in RCTL register.  [e1000_configure_rx()]

During the 'ifconfig ethN up' there is a window opened, starting in step B.2
where the receives are enabled up until they are disabled in step B.3, in
which the address of the receive descriptor dma page known by the hardware
emulator is still the previous one which was freed in step A.2. (In real
hardware the reset that is done by e1000_down() in step A.1 would have
cleared the previous values of RDBAH and RDBAL, but the emulator wasn't doing
a proper reset.) If this memory has been reallocated for some other use and
the hardware emulator feels so inclined, it will write to that former dma
page with predictably unpleasant results.

This problem doesn't appear to occur on bare-metal. Andy suspects that this
is because in the emulator, link-up is essentially instantaneous and traffic
can start flowing immediately. Whereas on bare-metal, link-up usually seems
to take at least a few milliseconds. And this might be enough to prevent
traffic from flowing into the device inside the window where E1000_RCTL_EN
is set before the proper address is set.

qemu-kvm's e1000 hardware emulator needs to be changed to do a proper reset.
But at the same time, it seems like a bad idea to enable dma operations
before the address of the memory to be involved has been made known.

This patch no longer enables receives in e1000_setup_rctl(), but leaves them
however they were. It only enables receives in e1000_configure_rx(), and
only after the dma address has been made known to the hardware.

There are two places where e1000_setup_rctl() gets called. The one in
e1000_configure() is followed immediately by a call to e1000_configure_rx(),
so there's really no change functionally (except for the removal of the
problem window. The other is in __e1000_shutdown() but is not followed by
a call to e1000_configure_rx(), so there is a change functionally. But
consider...

  C. An 'ifconfig ethN down' (just as described above in step A).

  D. A 'suspend' of the system, which (I'm assuming) will find its way into
     e1000_suspend() which calls __e1000_shutdown() resulting in:
     1. E1000_RCTL_EN being set in RCTL register.  [e1000_setup_rctl()]

And again we've re-opened the problem window for some unknown amount of time.