Fix module alt_instr KABI breakage
Patchwork-id: 40717
O-Subject: [RHEL6.2 PATCH BZ 737753 1/2 v2]: Fix module alt_instr KABI breakage
Bugzilla: 737753
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Jason Baron <jbaron@redhat.com>
RH-Acked-by: Kyle McMartin <kmcmartin@redhat.com>

*RHEL6 only*

RHEL6 commit 7c4233a9012846aa7998dad972fd1037d8ad6eb9 modified the cpuid
field in struct alt_instr from a u8 to a u16.  Some modules embed alt_instr
in the headers in order to export alternative instructions data to the kernel
and therefore older modules will no longer load properly.

ie) KABI is broken ... badly broken :(

The committed change results in an offset of 8 bits in the cpuid, instrlen, and
replacementlen fields of alt_instr and the offset must be applied to modules
that are from kernels that are 6.1 and older.

[v2]: The patch adds a new section to modules, .rheldata and initializes the
data in the rheldata struct.  The code then checks for the .rheldata section
because older modules will not have a .rheldata section and applies the
one byte shift to the data in alt_instr.

Successfully tested by gbeshers (on an SGI UV system where the bug was
originally reported) and by myself.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>