apic: ack all pending irqs when crashed/on kexec
Patchwork-id: 40738
O-Subject: [RHEL6.2 PATCH] x86, apic: ack all pending irqs when crashed/on kexec
Bugzilla: 704142
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Vivek Goyal <vgoyal@redhat.com>

BZ#704142
https://bugzilla.redhat.com/show_bug.cgi?id=704142

This is a backport of the upstream fix:
commit 8c3ba8d049247dc06b6dcee1711a11b26647aa44
Author: Kerstin Jonsson <kerstin.jonsson@ericsson.com>
Date:   Mon May 24 12:13:15 2010 -0700

Description:
A kernel panic "timer doesn't work through Interrupt-remapped IO-APIC"
sometimes occurs in 2nd kernel. This panic is caused by the pending
interrupt in local APIC which comes from 1st kernel. In the case of this bz,
IPMI interrupt(IRQ3) is left in APIC and it causes panic. This patch clears
such a pending interrupt in setup routine of local APIC. The following is
the description in the upstream patch, it well explains this problem.

  When the SMP kernel decides to crash_kexec() the local APICs may have
  pending interrupts in their vector tables.

  The setup routine for the local APIC has a deficient mechanism for
  clearing these interrupts, it only handles interrupts that has already
  been dispatched to the local core for servicing (the ISR register) safely,
  it doesn't consider lower prioritized queued interrupts stored in the IRR
  register.

  If you have more than one pending interrupt within the same 32 bit word in
  the LAPIC vector table registers you may find yourself entering the IO
  APIC setup with pending interrupts left in the LAPIC.  This is a situation
  for wich the IO APIC setup is not prepared.  Depending of what/which
  interrupt vector/vectors are stuck in the APIC tables your system may show
  various degrees of malfunctioning.  That was the reason why the
  check_timer() failed in our system, the timer interrupts was blocked by
  pending interrupts from the old kernel when routed trough the IO APIC.

Test status:
I tested on Fujitsu PRIMEQUEST(x86_64) and confirmed this problem is fixed
with this patch.

Successful brew scratch build
https://brewweb.devel.redhat.com/taskinfo?taskID=3641616

Please review and ACK.

Thanks,
Takao Indoh

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>