xfs: avoid synchronous transactions when deleting attr blocks
Patchwork-id: 40913
O-Subject: [RHEL6.2 PATCH] xfs: avoid synchronous transactions when deleting attr blocks
Bugzilla: 740312
RH-Acked-by: Christoph Hellwig <chellwig@redhat.com>
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

Upstream Commit: b73d8f7667aa82ece204a9a2e5467b54a8ecd059
RH BZ: 740312

This commit is cherry picked out of a series of cleanup patches that changes a
larger amount of code. Hence this commit is slightly larger than the upstream
commit, because it removes the dead code and comments that other upstream
commits removed or changed. We do not want to backport the entire cleanup
series just for this functionality late in the 6.2 cycle, so this is just the
fix with the assoicated removal of dead code that results from the fix.

Upstream commit message:

Currently xfs_attr_inactive causes a synchronous transactions if we are
removing a file that has any extents allocated to the attribute fork, and
thus makes XFS extremely slow at removing files with out of line extended
attributes. The code looks a like a relict from the days before the busy
extent list, but with the busy extent list we avoid reusing data and attr
extents that have been freed but not commited yet, so this code is just
as superflous as the synchronous transactions for data blocks.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Reported-by: Bernd Schubert <bernd.schubert@itwm.fraunhofer.de>
Reviewed-by: Dave Chinner <dchinner@redhat.com>
Signed-off-by: Alex Elder <aelder@sgi.com>
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>