GFS2: large file delete/unlink is slow
Patchwork-id: 40924
O-Subject: [RHEL6.2 Patch] bz 739987 - GFS2: large file delete/unlink is slow
Bugzilla: 739987
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

Hi,

This patch is a RHEL6.2 crosswrite of a performance patch
that was accepted and re-factored upstream.

This patch improves the performance of delete/unlink
operations in a GFS2 file system where the files are large
by adding a layer of metadata read-ahead for indirect blocks.
Mileage will vary, but on my system, deleting an 8.6G file
dropped from 22 seconds to about 4.5 seconds.

We decided to put it into 6.2 so that 6.2 is not slower
than 5.8, which I posted earlier this morning.

It was tested on RHEL6 system gfs-a16c-01.mpc.lab.eng.bos.redhat.com.

rhbz#739987

Regards,

Bob Peterson
Red Hat File Systems

Signed-off-by: Bob Peterson <rpeterso@redhat.com>
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>