netxen: Add firmware version 4.0.557[579].
Patchwork-id: 41048
O-Subject: [RHEL 6.2 PATCH] netxen: Add firmware version 4.0.557[579].
Bugzilla: 741776
RH-Acked-by: Andy Gospodarek <gospo@redhat.com>

Bugzilla
--------

Bug 741776 - [QLogic 6.2 Bug] netxen: Update firmware to 4.0.557 [579].

Brew Build
----------

3675246

Testing
-------

This patch was test built against the 2.6.32-203.el6 kernel.  The
firmware is released on ftp.qlogic.com and was tested internally by
QLogic.

Description
-----------

This patch satisfies the minimum requirement for the following HP
advisory:

http://h20000.www2.hp.com/bizsupport/TechSupport/Document.jsp?lang=en&cc=us&taskId=110&prodSeriesId=3913537&prodTypeId=329290&objectID=c02964542

I've updated WHENCE and the netxen_nic driver to pull in the correct
versioned firmware file.  The binary firmware was also converted to ihex
via the objcopy command.  Below is the patch (with diff stat) without
the firmware change.

Note that the firmware will come up as 4.0.579 but this is oddity with the
firmware versioning scheme...this firmware is the follow-on to 4.0.556 which
is mentioned in the advisory.

From: Chad Dupuis <cdupuis@redhat.com>

Subject: [PATCH] netxen: Add firmware version 4.0.579.
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>