acpi: Prevent acpiphp from deadlocking on PCI-to-PCI bridge remove
Patchwork-id: 41090
O-Subject: [RHEL6 PATCH BZ 732706 v2]: ACPI: Prevent acpiphp from deadlocking on
	PCI-to-PCI bridge remove
Bugzilla: 732706
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Myron Stowe <mstowe@redhat.com>

I originally submitted a patch to workaround this by pushing all Ejection
Requests and Device Checks onto the kacpi_hotplug queue.

http://marc.info/?l=linux-acpi&m=131678270930105&w=2

The patch is still insufficient in that Bus Checks also need to be added.

Rather than add all events to the hotplug queue, mjg suggested
that a better approach would be to modify the acpiphp driver such that only
acpiphp events would be added to the kacpi_hotplug queue.

It's a longer patch, but at least we maintain the benefit of having separate
queues for now.  This, of course, is still only a workaround the problem.
As Bjorn and mjg pointed out, we have to refactor a lot of this code to do the
right thing but at this point it is a better to have this code working.

P.
---8<---

The acpi core places all events on the kacpi_notify queue.  When the acpiphp
driver is loaded and a PCI card with a PCI-to-PCI bridge is removed the
following call sequence occurs:

cleanup_p2p_bridge()
	        -> cleanup_bridge()
	                -> acpi_remove_notify_handler()
	                        -> acpi_os_wait_events_complete()
	                                -> flush_workqueue(kacpi_notify_wq)

which is the queue we are currently executing on and the process will hang.

Move all hotplug acpiphp events onto the kacpi_hotplug workqueue.  In
handle_hotplug_event_bridge() and handle_hotplug_event_func() we can simply
push the rest of the work onto the kacpi_hotplug queue and then avoid the
deadlock.

Successfully tested by me on Hitachi platform.

Upstream: http://marc.info/?l=linux-pci&m=131725327802022&w=2
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>