bnx2fc: call ctlr_link_up only when the interface is enabled
Patchwork-id: 41354
O-Subject: [RHEL6.2 PATCH 5/9] bnx2fc: call ctlr_link_up only when the interface
	is enabled
Bugzilla: 740096
RH-Acked-by: Rob Evers <revers@redhat.com>

From: Bhanu Prakash Gollapudi <bprakash@broadcom.com>

Link may not be up when the driver receives ulp_start event, and hence
fcoe_ctlr_link_up is not called. Call fcoe_ctlr_link_up during
indicate_netevent only when the interface is enabled. (It has to be called when
enabled because that is an indication that the vlan discovery is completed).

This has been posted upstream:
http://www.spinics.net/lists/linux-scsi/msg54821.html

Signed-off-by: Bhanu Prakash Gollapudi <bprakash@broadcom.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>