Update hpet_next_event()
Patchwork-id: 41460
O-Subject: [RHEL6.3 PATCH BZ 728315 1/2]: Update hpet_next_event()
Bugzilla: 728315
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

The current hpet_next_event() code does the following:

cnt = hpet_readl(HPET_COUNTER);
cnt += (u32) delta;
hpet_writel(cnt, HPET_Tn_CMP(timer));

It is possible that an interrupt can occur between the read and write of the
HPET, and as a result the value in the counter (HPET_COUNTER) can be beyond
that being written to the comparator (HPET_Tn_CMP).

This results in timers being overdue -- as Uli notes @ 14.3MHz this results
in an ~5 min delay.

This can be resolved by using upstreams current implementation of the HPET
code in which a comparison is performed between the value of the counter and
the comparator.  If the counter is beyond the comparator, -ETIME will be
returned.

Patch 1/2 brings the HPET code up-to-date with linux-2.6 as of commit
f1c18071ad70e2a78ab31fc26a18fcfa954a05c6.  There are some additional benefits
of updating to linux-2.6, most notably we better manage systems in which there
may be a delay in the write of the comparator, however, my main goal is to
resolve the issue described above.

Successfully tested by customer, successfully compiled and tested by me on
a mix of AMD and Intel systems.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>