initialize min_delta_ns in one_hpet_msi_clockevent()
Patchwork-id: 41592
O-Subject: Re: [RHEL6.3 PATCH BZ 728315 2/2]: initialize min_delta_ns
	in	one_hpet_msi_clockevent()
Bugzilla: 728315

[RHEL6 PATCH BZ 728315 2/2]: initialize min_delta_ns in one_hpet_msi_clockevent()

Upstream handles the hpet code differently than that of RHEL.  In upstream
the code was changed to streamline the hpet initalization code and the
concept of clockevents was expanded.  This exact codepath no longer exists in
RHEL, however, initialization of the min_delta_ns is is done in linux-2.6 in
hpet_legacy_clockevent_register().

With the change to the hpet code in 1/2 we have to properly initalize the
minimum reprogramming delta appropriately.

RHEL6 only.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>