Lower USB storage settling delay to something more reasonable
Patchwork-id: 41611
O-Subject: [RHEL6 PATCH] Lower USB storage settling delay to something more
	reasonable
Bugzilla: 743959
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=743959

One of our customers was having trouble loading their usb driver disk
when booting from UEFI as opposed to a normal BIOS boot.  What appears
to happen is when booting from UEFI, the usb driver disk is detected
about a millisecond before udev is started.  This behaviour is different
when booting from BIOS by a swing of a millisecond in the other direction.

The effect of this difference is only understood by how anaconda and udev
work.  Anaconda starts udev as early as possible to catch all the hot
plug events (which is the case when booting from BIOS).  In the UEFI case,
because the event is generated before udev starts, udev doesn't see it.

The normal procedure for udev in this case is to run a 'udevadm trigger'
command immediately after udev is started.  This is how dracut does things
currently in RHEL-6.

However, anaconda has a special process and doesn't get around to running
this trigger until about 4-7 seconds later.  Anaconda runs the 'trigger' and
'settle' udevadm commands and then checks for a driver disk before moving
on.

This is where timing gets screwed up.  Because the driver disk is on a usb
stick, the usb_storage driver has a built-in 5 second settle period (that
udev has no clue about).  In the BIOS case this 5-second period is swallowed
up by the natural anaconda delay (4-7) before they check for the driver disk,
so everything just works (there is a couple of other small delays too for those
doing the math here).

In the UEFI case, because the event happens before udev starts, the usb_storage
driver doesn't get loaded until the 'trigger' event happens.  Because anaconda
runs the 'trigger', 'settle', driver disk so close together, the 5-second delay
doesn't finish until *after* anaconda determines there is no driver disk.

Eww.  Looking upstream, they reduce this ridiculous delay from 5 seconds to
1 second.  That fixes the problem on a couple of the customer's machines, but
they still have another machine with this problem.

Setting usb_storage.delay_use=0 on the command line fixes their problem, but
I am not about to deal with the fallout from that right now. So anaconda
has agreed to add in a 1 or 2 second delay before checking for a driver disk.
This helps eat up the usb delay and allows the customer's driver disk to work.

Regardless of the problem case, reducing this delay in general is good for
anyone who twiddles their thumbs after plugging in their usb drive and waits
for it to show up on their desktop.

This has been upstream for a year or so and it doesn't seem like anyone
has screamed yet, so I don't expect any regressions with this change.

commit a4a47bc03fe520e95e0c4212bf97c86545fb14f9
Author: Linus Torvalds <torvalds@linux-foundation.org>
Date:   Fri Feb 26 10:03:22 2010 -0800

    Lower USB storage settling delay to something more reasonable

    The five-second delay can be rather annoying, and makes the system
    appear much less responsive when you connect a USB drive.

    It's also not entirely clear that it is needed - the settling delay has
    at least historically been an issue on some Apple iPods, for example,
    and some devices have been reported to need even more than the old 5s
    delay.

    But before we penalize them all, let's see how bad it really is.  Some
    of the reasons for long delays seem to be actual historical kernel bugs
    that should probably never have been papered over with a delay in the
    first place (there's a Ubuntu bug report for 2.6.20 about a NULL pointer
    dereference unless 'delay_use' is 8 or more, for example).

    It also looks like some distros have already shipped with delay_use=0,
    so the five second default may well be totally historical.

    In other words: "Let's see if anybody screams".

    Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>