intel-iommu: IOTLB hang workaround
Patchwork-id: 41792
O-Subject: [PATCH 3/3] [RHEL6.2 PATCH] intel-iommu: IOTLB hang workaround
Bugzilla: 728476
RH-Acked-by: Adam Jackson <ajax@redhat.com>

bz: 728476
upstream: on the way for 3.2

To work around a hardware issue, we have to submit IOTLB flushes while
the graphics engine is idle. The graphics driver will go to great
lengths to ensure that it gets that right on the affected chipset(s)...
so let's not screw it over by deferring the unmap and doing it later.
That wouldn't be very helpful.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>
Signed-off-by: Ben Widawsky <ben@bwidawsk.net>
Signed-off-by: Dave Airlie <airlied@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>