bnx2i: Fixed the endian on TTT for NOP out transmission
Patchwork-id: 41858
O-Subject: [RHEL6.2 PATCH 1/1] bnx2i: Fixed the endian on TTT for NOP out
	transmission
Bugzilla: 745676
RH-Acked-by: Rob Evers <revers@redhat.com>
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>

From: Eddie Wai <eddie.wai@broadcom.com>

This is for bz 745676.

This patch is low risk. It only affects bnx2i users, but it
occurs with a popular target from Dell. Without the patch
there will be a perforamnce drop at best, and at the worse
IO could be failed to the application when it should have
succeeded.

I have tested the patch by forcing the initiator to send nops
(iscsi pings) to stress that code. And, Broadcom has confirmed
it fixes the issue.

>From upstream commit bc41435d7b1f6403b1454fe135e17406f3421b3f
in the scsi maintainer's scsi-rc-fixes tree
git://git.kernel.org/pub/scm/linux/kernel/git/jejb/scsi-rc-fixes-2.6.git:

The iscsi_nopout task's TTT is defined as __be32 while the DMA
memory to the chip is CPU specific.  This creates a problem for
unsolicited NOP-In responses where the TTT is not the RESERVED
tag of 0xFFs.  This patch adds a call to be32_to_cpu for the TTT
specified.

Signed-off-by: Eddie Wai <eddie.wai@broadcom.com>
Signed-off-by: James Bottomley <JBottomley@Parallels.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>